+++ management/connection.go
@@ -1142,6 +1142,9 @@ type ConnectionOptionsOAuth2 struct {
 
 	// UpstreamParams specifies upstream parameters.
 	UpstreamParams map[string]interface{} `json:"upstream_params,omitempty"`
+
+	// CustomHeaders specifies custom headers.
+	CustomHeaders map[string]string `json:"custom_headers,omitempty"`
 }
 
 // UnmarshalJSON implements the json.Unmarshaler interface for ConnectionOptionsOAuth2.